+++ backend/src/main/java/cloudcode/guestbook/backend/UserController.java
@@ -1,0 +1,52 @@
+package cloudcode.guestbook.backend;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
+
+/**
+ * defines the REST endpoints managed by the server.
+ */
+@RestController
+public class UserController {
+
+  @Autowired
+  private UserRepository userRepository;
+
+  @Autowired
+  private CustomUserDetailsService userService;
+  
+  @PostMapping("/login")
+  public final UserResponse login(@RequestBody User user) {
+    User userByName = userRepository.findByUsername(user.getUsername());
+    User match = userRepository.findByUsernameAndPassword(
+      user.getUsername(),
+      user.getPassword()
+    );
+
+    if (userByName == null) {
+      return new UserResponse(false, "No Account with Username");
+    } else if (match == null) {
+      return new UserResponse(false, "Incorrect Password");
+    } else {
+      return new UserResponse(true, null);
+    }
+  }
+
+
+  @PostMapping("/signup")
+  public final UserResponse addUser(@RequestBody User user) {
+    Boolean emailExists = userService.findUserByEmail(user.getEmail()) != null;
+    Boolean usernameExists =
+      userService.findUserByUsername(user.getUsername()) != null;
+    if (emailExists) {
+      return new UserResponse(false, "Email already registered");
+    } else if (usernameExists) {
+      return new UserResponse(false, "Username already registered");
+    } else {
+      userRepository.save(user);
+      return new UserResponse(true, null);
+    }
+  }
+}
